requirements.txt
pytorch
torchvision
torchaudio
cudatoolkit=11.6
pytorch-lightning==1.7.1
h5py==3.7.0
matplotlib==3.5.3
numpy==1.23.1
pandas==1.4.3
zarr==2.12.0
